removing warnings
git-svn-id: file:///home/behr_mi/git/sumo_synched/trunk@8674 afbd958f-9f77-42d5-a016-97a22340ccf4

+++ sumo/src/utils/gui/settings/GUICompleteSchemeStorage.cpp
@@ -148,7 +148,7 @@ GUICompleteSchemeStorage::init(FXApp *app) throw() {
         vs.minVehicleSize = 0;
         gSchemeStorage.add(vs);
     }
-    myNumInitialSettings = mySortedSchemeNames.size();
+    myNumInitialSettings = (unsigned int) mySortedSchemeNames.size();
     // add saved settings
     int noSaved = app->reg().readIntEntry("VisualizationSettings", "settingNo", 0);
     for (int i=0; i<noSaved; ++i) {
@@ -170,7 +170,7 @@ GUICompleteSchemeStorage::init(FXApp *app) throw() {
                     break;
                 }
                 content += part;
-                xmlSize -= part.size();
+                xmlSize -= (int) part.size();
                 index++;
             }
             if (content != "" && xmlSize == 0) {
